RCAagent/specializedAgents/DiagnosticAnalyst/pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src"]

[project]
name = "DiagnosticAnalyst"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "bedrock-agentcore >= 1.0.3",
    "google-auth >= 2.0.0",
    "gspread >= 6.0.0",
    "langchain >= 1.0.3",
    "langchain-mcp-adapters >= 0.1.11",
    "langchain_aws >= 1.0.0",
    "langgraph >= 1.0.2",
    "mcp >= 1.19.0",
    "pytest >= 7.0.0",
    "pytest-asyncio >= 0.21.0",
    "requests >= 2.32.5",
    "tiktoken==0.11.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["test"]